Add a layout test to test resuming HTMLModelElement after exiting the back/forward cache https://bugs.webkit.org/show_bug.cgi?id=292813 rdar://151048734
Reviewed by Mike Wyrzykowski.

This test page has an autoplay <model> element that should start
playing on load. After navigating to a different page, it would
temporarily pause the animation while suspended, and resume playing
after the element is resumed when the page is navigated back to.

* LayoutTests/model-element/model-element-suspend-resume-expected.txt: Added.
* LayoutTests/model-element/model-element-suspend-resume.html: Added.
* LayoutTests/model-element/resources/go-back-on-load.html: Added.

Canonical link: https://commits.webkit.org/294751@main