REGRESSION(288714@main): CanvasRenderingContext2D font with normal line height fails to parse https://bugs.webkit.org/show_bug.cgi?id=291667 rdar://149459999
Reviewed by Darin Adler.

Fixes regression in 288714@main that failed to parse the keyword `normal`
when parsing strings passed to `CanvasRenderingContext2D.font`.

The issue was due failing to include CSSPropertyParserConsumer+KeywordDefinitions.h
in CSSPropertyParserConsumer+Font.cpp, which meant that the `MetaConsumer` didn't have
a consumer for keyword values when being invoked in `consumeLineHeightUnresolved`.

To ensure this kind of issue can't happen in the future, we now statically assert
that all types passed to `MetaConsumer have `ConsumerDefinition` specializations
defined. Confirmed this works by just adding the assertion, which immediately caused
the build to fail pointing directly to `consumeLineHeightUnresolved`.

* LayoutTests/fast/canvas/font-line-height-normal-expected.txt: Added.
* LayoutTests/fast/canvas/font-line-height-normal.html: Added.
* Source/WebCore/css/parser/CSSPropertyParserConsumer+Font.cpp:
* Source/WebCore/css/parser/CSSPropertyParserConsumer+MetaConsumer.h:
* Source/WebCore/css/parser/CSSPropertyParserConsumer+MetaConsumerDefinitions.h:

Canonical link: https://commits.webkit.org/294767@main