REGRESSION ( 294005@main): [ iOS Debug ] ASSERTION FAILED: !isInGPUProcess() system-preview/badge.html is a cosnsitent crash https://bugs.webkit.org/show_bug.cgi?id=292758 rdar://151044237
Reviewed by Antoine Quint.

During the review of 294005@main it was thought this
code never ran in the GPU process, however there is a code path
which was caught by system-preview/badge.html.

Conditionalize the code path and undo skipping of the test.

* Source/WebCore/Modules/system-preview/ARKitBadgeSystemImage.mm:
(WebCore::ARKitBadgeSystemImage::draw const):
Since this may run in the GPU or web process, only use the IOSurface
path when running in the GPU process.

* LayoutTests/platform/ios/TestExpectations:
Revert skipping this test.

Canonical link: https://commits.webkit.org/294793@main