[scroll-animations] Crash under AnimationEffectTiming::getBasicTiming https://bugs.webkit.org/show_bug.cgi?id=292742 rdar://149951894
Reviewed by Antoine Quint.

Under AnimationEffectTiming::getBasicTiming, there is an assertion that if we have a resolved
start time for an animation, the animation's timeline should have a resolved current time. It
is fairly easy to get into the case where this doesn't hold true, for a scroll driven animation
where the scroll container becomes non-scrollable after the start time has been resolved. My
original solution was to make the start time unresolved in WebAnimation::autoAlignStartTime
if we saw that the timeline's time was unresolved, but after https://commits.webkit.org/294121@main
changed the timing of when we update the timeline's cached current time, it is possible for the
cached time to become unresolved before we can run WebAnimation::autoAlignStartTime. A simpler
solution is to remove the assert and check that both the start time and the timeline's current
time are resolved.

* LayoutTests/TestExpectations:
* LayoutTests/animations/sda-crash-expected.txt: Added.
* LayoutTests/animations/sda-crash.html: Added.
* LayoutTests/platform/mac-wk2/TestExpectations:
* Source/WebCore/animation/AnimationEffectTiming.cpp:
(WebCore::AnimationEffectTiming::getBasicTiming const):

Canonical link: https://commits.webkit.org/294813@main