[PGO] Work around memory corruption in instrumentation runtime https://bugs.webkit.org/show_bug.cgi?id=292899 rdar://151191066
Reviewed by Alexey Proskuryakov.

When building with instrumentation, move the __llvm_prf_cnts section to
its own segment. This section is the table of function counters
generated for a binary. At runtime, the compiler-rt library mmaps this
section to the .profdata file for that framework and process.

However, it rounds up to the nearest page boundary, which may be past
the end of the section. When the counts table is not itself
page-sized, the call to mmap might write over another section in __DATA,
such as ObjC runtime data (rdar://151083138).

We encounter this as `NSProtocolFromString(@"WKObject")` returning
garbage data in some instrumented builds of WebKit. But it could happen
to anything in __DATA. By moving __llvm_prf_cnts to its own segment, it
is guaranteed to be the only section in its page and can mmap freely.

* Configurations/CommonBase.xcconfig:

Canonical link: https://commits.webkit.org/294823@main