[iOS] sciencedirect.com login page input field has unexpected rounded bottom border https://bugs.webkit.org/show_bug.cgi?id=292722 rdar://146849231
Reviewed by Abrar Rahman Protyasha and Richard Robinson.

The input element on sciencedirect.com's login page is styled to remove the
border on the left, right, and top edges. Additionally, it changes the width
and color of the bottom border.

This author styling should "devolve" the widget, as "border" is an
appearance-disabling property. As a result, the input element should lose its
native appearance, and should not have a border-radius. However, on iOS,
border-radius is specified to be 5px in the UA stylesheet. Consequently, the
platform-specific border-radius is always applied to the control, regardless of
whether or not native appearance is used. This behavior differs from macOS
and other browsers, which do not apply a border-radius for inputs with
non-native appearance.

Fix by removing the platform-specific border radius from the UA stylesheet, and
into style adjustment for native appearances that require a 5px border radius.

Rebaseline tests that modify the background/border on input and textarea elements,
to account for the loss of the border-radius.

* LayoutTests/fast/forms/input-text-default-border-radius-with-author-styled-border-expected.txt: Added.
* LayoutTests/fast/forms/input-text-default-border-radius-with-author-styled-border.html: Added.
* LayoutTests/interaction-region/text-controls-expected.txt:
* LayoutTests/interaction-region/text-input-focused-edited-empty-expected.txt:
* LayoutTests/interaction-region/text-input-focused-expected.txt:
* LayoutTests/interaction-region/textarea-focused-edited-empty-expected.txt:
* LayoutTests/interaction-region/textarea-focused-expected.txt:
* LayoutTests/platform/ios/TestExpectations:

Mark more `compute-kind-widget-generated` as failing. These failures are not
entirely real, since setting `border-radius` does correctly disable native
appearance. However, these generated tests expect that the computed style for
`appearance: auto` and `appearance: none` are equivalent. `appearance: auto`
for the controls affected by this change has `border-radius: 5px`, while
`appearance: none` has `border-radius: 0px`.

While it is a long term goal to make the computed styles equivalent, that is
generally not the case on Cocoa platforms today. This is evidenced by the existing
`compute-kind-widget-generated` failures.

* LayoutTests/platform/ios/compositing/overflow/textarea-scroll-touch-expected.txt:
* LayoutTests/platform/ios/fast/scrolling/ios/textarea-scroll-touch-expected.txt:
* Source/WebCore/css/html.css:
* Source/WebCore/page/InteractionRegion.cpp:
(WebCore::canTweakShapeForStyle):
(WebCore::interactionRegionForRenderedRegion):

Exclude text inputs and textareas from being eligible for bounds extension.
Previously, they were ineligible as a consequence of `border-radius: 5px;`
in the UA stylesheet, and a check for `RenderStyle::hasExplicitlySetBorderRadius()`.
However, now that the border-radius is set during style adjustment,
`hasExplicitlySetBorderRadius` is (correctly) false.

Without this change, `LayoutTests/interaction-region/text-input.html` would fail,
as a border-radius of 8px (InteractionRegionMinimumCornerRadius) would be used for
the interaction region for text inputs, even though the used border-radius is 5px.

* Source/WebCore/rendering/ios/RenderThemeIOS.mm:
(WebCore::applyCommonNonCapsuleBorderRadiusToStyle):
(WebCore::RenderThemeIOS::adjustTextFieldStyle const):
(WebCore::RenderThemeIOS::adjustTextAreaStyle const):
(WebCore::adjustInputElementButtonStyle):

Canonical link: https://commits.webkit.org/294843@main