Remove CSSOMViewSmoothScrollingEnabled preference https://bugs.webkit.org/show_bug.cgi?id=292867
Reviewed by Antti Koivisto.

This has been enabled on main for over two years.

This is also in a weird state: enabled for WebKit1, but disabled for
testing on WebKit1. By also enabling it for testing on WebKit1 we can
reduce the number of WebKit1-specific expectations and more importantly
ensure it is actually tested.

Canonical link: https://commits.webkit.org/294846@main