Attempting to modify selection while zoomed in results in fast, unexpected scrolling. https://bugs.webkit.org/show_bug.cgi?id=292897 rdar://150884677
Reviewed by Aditya Keerthi.

With the change in 290497@main I changed the code to have a static offset
that didn't change with zoom. Before it was taking the minimum of a static
offset and the delta of how far the point had traveled, which was almost
always small, so it was always the small number, and scrolling was slow
but also wasn't affected by zooming. We just need to scale the static
offsets and speed by the current zoom factor to keep a consistent experience
at any zoom level.

* Source/WebCore/page/ios/EventHandlerIOS.mm:
(WebCore::adjustAutoscrollDestinationForInsetEdges):
(WebCore::EventHandler::targetPositionInWindowForSelectionAutoscroll const):

Canonical link: https://commits.webkit.org/294874@main