premierleague.com: scrubbing through video content invokes text selection UI https://bugs.webkit.org/show_bug.cgi?id=292998 rdar://136791737
Reviewed by Abrar Rahman Protyasha.

Augment an existing quirk, `shouldAvoidStartingSelectionOnMouseDownOverPointerCursor`, such that it
prevents text selection over all elements with `cursor: pointer;`, and extend it so that it applies
to premierleague.com as well.

* Source/WebCore/page/Quirks.cpp:
(WebCore::Quirks::shouldAvoidStartingSelectionOnMouseDownOverPointerCursor const):

Pull this quirk out of `PLATFORM(MAC)`, so that it can apply to iOS as well when using a trackpad to
select text.

(WebCore::handleSpotifyQuirks):
* Source/WebCore/page/Quirks.h:
* Source/WebCore/page/QuirksData.h:
* Source/WebKit/Shared/ios/InteractionInformationAtPosition.h:
* Source/WebKit/Shared/ios/InteractionInformationAtPosition.serialization.in:

Rename `UnselectableDueToUserSelectNone` to `UnselectableDueToUserSelectNoneOrQuirk`, and set the
flag when we're interacting with a target node where the quirk is active.

* Source/WebKit/UIProcess/ios/WKContentViewInteraction.mm:
(-[WKContentView textInteractionGesture:shouldBeginAtPoint:]):
* Source/WebKit/WebProcess/WebPage/ios/WebPageIOS.mm:
(WebKit::selectionPositionInformation):

Canonical link: https://commits.webkit.org/294931@main