BrowsingContextGroup should only be shared from opener to opened WKWebViews https://bugs.webkit.org/show_bug.cgi?id=293027 rdar://151345757
Reviewed by Sihui Liu.

When WKUIDelegate.createWebViewWithConfiguration is called, it gives the application
a WKWebViewConfiguration to use to make a WKWebView and put it in the same
BrowsingContextGroup.  If a WKWebViewConfiguration is reused, we need to not
use the same BrowsingContextGroup, which is web compatible because there is no
opener relationship to maintain in that case.  To implement this, we consume
the configuration's opener info when making a WebPageProxy so the configuration
can be further copied and reused without adding more pages to the BrowsingContextGroup.

An app can copy the opener's configuration and use that configuration to load
the opener site in an auxiliary WKWebView before responding to
WKUIDelegate.createWebViewWithConfiguration. When this happens with site isolation off,
the auxiliary WKWebView loads in its own process.  When this happens with site isolation
on before this PR, it would try to load in the process that was waiting for the reply
to Messages::WebPageProxy::CreateNewPage, which caused a deadlock.  This solution
avoids the deadlock, increases compatibility, and allows the authentication to proceed.

This fixes the last API test in the SOAuthorizationPopUp group with site isolation on.

* Source/WebKit/UIProcess/API/APIPageConfiguration.cpp:
(API::PageConfiguration::Data::Data):
(API::PageConfiguration::openerInfo const const):
(API::PageConfiguration::setOpenerInfo):
(API::PageConfiguration::consumeOpenerInfo):
(API::PageConfiguration::Data::createBrowsingContextGroup): Deleted.
(API::PageConfiguration::browsingContextGroup const): Deleted.
(API::PageConfiguration::setBrowsingContextGroup): Deleted.
* Source/WebKit/UIProcess/API/APIPageConfiguration.h:
(API::PageConfiguration::Data::initializer):
* Source/WebKit/UIProcess/Cocoa/UIDelegate.mm:
(WebKit::UIDelegate::UIClient::createNewPage):
* Source/WebKit/UIProcess/WebPageProxy.cpp:
(WebKit::m_pageForTesting):
(WebKit::WebPageProxy::initializeWebPage):
(WebKit::WebPageProxy::createNewPage):
* Source/WebKit/UIProcess/WebPageProxy.h:
* Tools/TestWebKitAPI/Tests/WebKitCocoa/SiteIsolation.mm:
(TestWebKitAPI::TEST(SiteIsolation, LoadDuringOpen)):
(TestWebKitAPI::TEST(SiteIsolation, ChildNavigatingToDomainLoadedOnADifferentPage)):
(TestWebKitAPI::TEST(SiteIsolation, ReuseConfigurationLoadHTMLString)):

Canonical link: https://commits.webkit.org/294968@main