FontCascade::width accepts a StringView rdar://151404312 https://bugs.webkit.org/show_bug.cgi?id=293060
Reviewed by Brent Fulgham.

There are a few cases where we instantiate a TextRun from a String because we want
to measure FontCascade::width and this function works with runs.

We are providing a wrapper that instantiates the TextRun from a StringView so
we don't have to do that on every call.

* Source/WebCore/SaferCPPExpectations/UncheckedCallArgsCheckerExpectations:
* Source/WebCore/inspector/InspectorOverlayLabel.cpp:
(WebCore::InspectorOverlayLabel::expectedSize):
* Source/WebCore/layout/formattingContexts/inline/display/InlineDisplayLineBuilder.cpp:
(WebCore::Layout::trailingEllipsisVisualRectAfterTruncation):
* Source/WebCore/layout/formattingContexts/inline/text/TextUtil.cpp:
(WebCore::Layout::TextUtil::hyphenWidth):
* Source/WebCore/platform/graphics/FontCascade.cpp:
(WebCore::FontCascade::width const):
* Source/WebCore/platform/graphics/FontCascade.h:
* Source/WebKit/WebProcess/WebCoreSupport/win/WebPopupMenuWin.cpp:
(WebKit::WebPopupMenu::setUpPlatformData):

Canonical link: https://commits.webkit.org/294986@main