WebProcess should not forward logging happening in realtime threads rdar://151379286 https://bugs.webkit.org/show_bug.cgi?id=293044
Reviewed by Per Arne Vollan.

We do not want WebProcess::registerLogHook lambda to execute in realtime threads as it is allocating memory and holiding locks.
To prevent this, we add a helper routine that allows checking whether the current thread is realtime.
If so, we exit early in WebProcess::registerLogHook lambda.

* Source/WTF/wtf/Threading.cpp:
(WTF::Thread::create):
(WTF::Thread::currentThreadIsRealtime):
* Source/WTF/wtf/Threading.h:
* Source/WebKit/WebProcess/cocoa/WebProcessCocoa.mm:
(WebKit::WebProcess::registerLogHook):

Canonical link: https://commits.webkit.org/294998@main