[scroll-animations] debug assertion hit under `JSCell::classInfo()` during document teardown when the last Document reference is dereferenced https://bugs.webkit.org/show_bug.cgi?id=293108 rdar://151306277
Reviewed by Ryosuke Niwa.

When the last reference to a `Document` is dereferenced, document teardown occurs and we end up calling `WebAnimation::remove()`
for all animations still attached to the document under `AnimationTimeline::detachFromDocument()`. In `WebAnimation::remove()`
we ensure we clear strong references to the effect and timeline.

In 235621@main, we opted to use `setTimeline()` instead of `setTimelineInternal()` to clear the timeline reference, which would
reset some playback-related state and avoid a crash if using the DOM API `updatePlaybackRate()`. But with the advent of Scroll-driven
Animations, this is the source of a new issue where, if a Scroll-driven Animation has a pending pause task, then the "setting
the timeline" [1] procedure will resolve the ready promise which should not happen during document teardown caused by a dereference
of a final Document reference. Specifically we end up hitting `ASSERT(vm().heap.mutatorState() != MutatorState::Sweeping)` under
`JSCell::classInfo()`.

To fix this specific issue, we go back to using `setTimelineInternal()` in `WebAnimation::remove()` which will simply ensure we
reset the strong reference to the current timeline, without any further processing. As for the issue initially addressed in
235621@main, we reset both `m_holdTime` and `m_startTime` such that the playback state is reset appropriately.

[0] https://drafts.csswg.org/web-animations-1/#pending-pause-task
[1] https://drafts.csswg.org/web-animations-2/#setting-the-timeline

* LayoutTests/TestExpectations:
* LayoutTests/webanimations/document-teardown-due-to-last-reference-crash-expected.txt: Added.
* LayoutTests/webanimations/document-teardown-due-to-last-reference-crash.html: Added.
* Source/WebCore/animation/WebAnimation.cpp:
(WebCore::WebAnimation::remove):

Canonical link: https://commits.webkit.org/295009@main